backend/requirements.txt
pillow==5.4.1
django==2.1.5
django-tenants==2.2.3
channels==2.1.7
channels-redis==2.3.3
django-cors-headers==2.4.0
djangorestframework==3.9.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.9
